Update GCSE equivalency testing wording
Testing showed that some providers were reluctant to tick the subject boxes as they didn't run the equivalency tests themselves, but instead required candidates to use a third party.

This tweaked wording hopefully is generic enough to cover both cases, and the free text field uses the hint text to encourage providers to explain how they offer or require the tests to be taken. This is also now non-optional.