Remove actions banners
These were often missed or ignored in user research, and so don’t seem particularly helpful.

The "Prepare for next cycle" banner is also potentially misleading, as it only highlights the NEW fields, but providers also need to review and update relevant existing fields (eg if fees have changed).

There is minimal downside to removing it, as if providers try to publish a course without providing the required information, they will see an error summary which highlights missing fields that prevents them from publishing. This was much more widely used in user research.